Removed Errors::TwoOperatorsInARow. Its previous use was semi-replaced by IncompleteOperation. This is to allow expressions such as a+-b.
While I personally put signed expressions within parentheses
(such as "a+(-b)"), it is a matter a personal preference and
both forms can only lead to one interpretation.
The above expression must therefore be considered legal.

Furthermore, expressions like "a+++b" and "a---b" (or other
permutations) are now allowed. Although they can be seen as
unusual, they are not ambiguous in the context of arithmetic.

Finally, validation is mainly supposed to prevent errors
during parsing. As such, all expressions which can be
parsed correctly should be allowed.